Create peopleList
This week you will be building on the Week 7's assignment.  The list of owners, managers and employees will be stored in a class.  Provide functions that will display all employees, or only 1 category.  Also provide a function that will sort the list by name alphabetically.  You will also need functions to add and remove people from the list.  Provide a menu similar to the one below to access the functionality:

1. Add person

2. Display list

3. Sort list

4. Quit



Option 1 should display a menu allowing the user to choose the type of person they are adding:

1. Owner

2. Manager

3. Laborer

Option 2 should display a menu allowing the user to choose the category of person to display

1. All

2. Owners

3. Managers

4. Laborers.