Create DoublyLinkedList
In the video examples we developed a linked list class which consisted on a node, made up of data and a pointer to the next node in the linked list.  This type of node works well for walking through a linked list from the beginning to the end.  One disadvantage is that it is inefficient to work through the linked list in reverse (from the end to the beginning).

 

A doubly linked list consists of nodes that are made up of a data part and one pointer that points to the next node and another pointer that points to the previous node in the list.

 

Note that in the example program of the linked list in the videos, there would need to be a pointer to the first node in the list, as well as the last.

 

The advantage is that the list can be traversed in both directions efficiently.  The disadvantage is that it will complicate the functions to add and remove nodes.

 

Create a doubly linked list class that will store <double> values.  The class should have the following:

Constructor

Destructor

Insert function

Remove function

Accessor function

Mutator function

 

Use this class in a program that will give a menu similar to below:

Add Value
Delete Value
Display List (forward)
Display List (backward)
Quit