Create Final
For the final project you will revisit the "Business" assignment from Week 7.

All requirements from that assignment apply to the final project with the exception that you are not required to use a vector.  You may use a vector or any other container that we've discussed in class.

In addition to the menu option to add a person to the container and display all person information in the list, you must also provide menu options to do the following:

1. Find a person by name then print out all information stored about that person

2. Find and remove a person by name

3. Calculate and display the average amount earned for each of the 3 categories of people

Finally - when the user chooses the option to quit - all information from the container will be written to a file.  When the program starts, that file will be read and the data preloaded into the container.